--- boost/__init__.py
@@ -1,2 +1,0 @@
-from __future__ import absolute_import, division, print_function
-import libtbx.forward_compatibility
