+++ boost/std_pair.py
@@ -1,4 +1,11 @@
 from __future__ import absolute_import, division, print_function
-import boost.python
-ext = boost.python.import_ext("std_pair_ext")
-from std_pair_ext import *
+from boost_adaptbx.std_pair import *
+import warnings
+
+warnings.warn(
+  "importing from boost.std_pair is deprecated; this module will be removed shortly. "
+  "import from boost_adaptbx.std_pair instead. "
+  "Please see https://github.com/cctbx/cctbx_project/issues/458 for more information.",
+  FutureWarning,
+  stacklevel=2
+)
