+++ boost_adaptbx/__init__.py
@@ -1,0 +1,2 @@
+from __future__ import absolute_import, division, print_function
+import libtbx.forward_compatibility
