+++ boost_adaptbx/std_pair.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, division, print_function
+import boost.python
+ext = boost.python.import_ext("std_pair_ext")
+from std_pair_ext import *
