+++ boost_adaptbx/tuple.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, division, print_function
+import boost.python
+ext = boost.python.import_ext("boost_tuple_ext")
+from boost_tuple_ext import *
